refactor: lotto 생성자에서 검증
