+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,13 +1,16 @@
 package lotto.domain.lotto;
 
-import java.util.Collections;
-import java.util.List;
+import lotto.message.ErrorMessage;
+
+import java.util.*;
 
 public class Lotto {
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
         validate(numbers);
+        validateDuplicateNumbers(numbers);
+        validateSortedAscending(numbers);
         this.numbers = numbers;
     }
 
@@ -18,15 +21,39 @@ public class Lotto {
     }
 
     // TODO: 추가 기능 구현
+    public void validateDuplicateNumbers(List<Integer> numbers) {
+        Set<Integer> checkNumbers = new HashSet<>();
+
+        for (int number : numbers) {
+            if(!checkNumbers.add(number)) {
+                throw new IllegalArgumentException(ErrorMessage.DUPLICATE_LOTTO_NUMBER.getMessage());
+            }
+        }
+    }
+
+    private void validateSortedAscending(List<Integer> numbers) {
+        List<Integer> sortedNumber = getSortedNumber(numbers);
+        if (!sortedNumber.equals(numbers)) {
+            throw new IllegalArgumentException(ErrorMessage.LOTTO_NUMBERS_NOT_SORTED.getMessage());
+        }
+    }
+
+    private List<Integer> getSortedNumber(List<Integer> numbers) {
+        List<Integer> sortedLotto = new ArrayList<>(numbers);
+        Collections.sort(sortedLotto);
+
+        return sortedLotto;
+    }
+
     public String lottoNumbersAsString() {
         return SupportedLotto.getLottoNumbersAsString(this);
     }
 
     public int lottoMatchValue(Lotto winningLotto) {
-        List<Integer> winningNumbers = winningLotto.numbers;
+        final List<Integer> winningNumbers = winningLotto.numbers;
 
         return (int) numbers.stream()
-                .filter(number -> winningNumbers.contains(number))
+                .filter(winningNumbers::contains)
                 .count();
     }
