+++ src/main/java/lotto/view/input/LottoBuyInputView.java
@@ -58,10 +58,8 @@ public class LottoBuyInputView extends Input{
     }
 
     private void validateWiningLottoNumbers(List<Integer> numbers) {
-        inputValidator.checkForDuplicateNumbers(numbers);
         inputValidator.validateLottoSize(numbers);
         inputValidator.validateLottoNumberRange(numbers);
-        inputValidator.validateSortedAscending(numbers);
     }
 
     public int requestBonusLottoNumber() {
