+++ src/test/java/lotto/validator/InputValidatorTest.java
@@ -48,17 +48,6 @@ class InputValidatorTest {
     }
 
     @Test
-    @DisplayName("번호가 중복되면 예외가 발생한다.")
-    public void checkForDuplicateNumbers() {
-        // given
-        List<Integer> numbers = List.of(1, 3, 3, 4, 5, 6);
-        // when // then
-        assertThatThrownBy(() -> inputValidator.checkForDuplicateNumbers(numbers))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(ErrorMessage.DUPLICATE_LOTTO_NUMBER.getMessage());
-    }
-
-    @Test
     @DisplayName("번호들이 1~45사이의 숫자가 아니면 예외가 발생한다.")
     public void validateLottoNumberRangeList() {
         // given
@@ -78,17 +67,6 @@ class InputValidatorTest {
         assertThatThrownBy(() -> inputValidator.validateLottoSize(numbers))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.INVALID_LOTTO_NUMBER_COUNT.getMessage());
-    }
-
-    @Test
-    @DisplayName("로또 번호가 오름차순으로 정렬되어 있지 않으면 예외가 발생한다.")
-    public void validateSortedAscending() {
-        // given
-        List<Integer> numbers = List.of(1, 2, 4, 3, 5, 6);
-        // when // then
-        assertThatThrownBy(() -> inputValidator.validateSortedAscending(numbers))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(ErrorMessage.LOTTO_NUMBERS_NOT_SORTED.getMessage());
     }
 
     @Test
